Gonit should set a different PPID for processes started by it
We are finding issues with Apache and the way it is stopped.
When doing that, it is sending a kill signal to all parent
processes, including Gonit, which is then shutdown with all
that it implies: Gonit is not managing processes anymore, and
"gonit start" does not work until you start its daemon.